images/ai-agent/raphael/backend/discord_bot.py
import discord
from discord.ext import commands

class RaphaelBot(commands.Bot):
    def __init__(self, *args, **kwargs):
        intents = discord.Intents.default()
        intents.message_content = True
        super().__init__(*args, intents=intents, **kwargs)

    async def on_ready(self):
        print(f'Raphael has awakened as {self.user} (ID: {self.user.id})')
        print('------')

    async def handle_alert(self, alert_data: dict):
        """
        Process incoming alert and notify Discord channel
        """
        # TODO: Implement rich embed generation
        channel_id = int(os.getenv("DISCORD_CHANNEL_ID", "0"))
        channel = self.get_channel(channel_id)
        if channel:
            embed = discord.Embed(
                title="🚨 Incident Detected",
                description=alert_data.get("message", "No message provided"),
                color=discord.Color.red()
            )
            await channel.send(embed=embed)

    @commands.command()
    async def status(self, ctx):
        await ctx.send("Raphael is monitoring the cluster. All systems operational.")

    @commands.command()
    async def report(self, ctx):
        # TODO: Implement financial report generation
        await ctx.send("Generating Financial Efficiency Report... (Coming Soon)")
