images/ai-agent/raphael/backend/requirements.txt
fastapi
uvicorn
discord.py
kubernetes
qdrant-client
redis
requests
GitPython
prometheus-api-client
python-dotenv
aiohttp
